+++ internal/datasource_device_ap_stats/sdk_to_terraform_last_trouble.go
@@ -20,7 +20,7 @@ func lastTroubleSdkToTerraform(ctx context.Context, diags *diag.Diagnostics, d *
 		code = types.StringValue(*d.Code)
 	}
 	if d.Timestamp != nil {
-		timestamp = types.Float64Value(float64(*d.Timestamp))
+		timestamp = types.Float64Value(*d.Timestamp)
 	}
 
 	dataMapValue := map[string]attr.Value{
